- Epidemic protocol seems to be working correctly with 2 nodes running locally

+++ src/com/g8A/CPEN431/A7/server/distribution/VirtualNode.java
@@ -6,9 +6,9 @@ public class VirtualNode implements Node {
 
     private final ByteString pNode;
     private int pNodeId;
-    private final long vIndex;
+    private final int vIndex;
 
-    public VirtualNode(ByteString pNode, int pNodeId, long vIndex) {
+    public VirtualNode(ByteString pNode, int pNodeId, int vIndex) {
         this.pNode = pNode;
         this.pNodeId = pNodeId;
         this.vIndex = vIndex;
